requirements.txt
anyio==3.7.1
bcrypt==4.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.0.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
Jinja2==3.1.2
MarkupSafe==3.0.2
pyasn1==0.6.1
pydantic>=2.0.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn[standard]==0.23.2
websockets==11.0.3